+++ frontend/src/components/Home/Home.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import IconSelector from './IconSelector';
 import SearchBar from './SearchBar';
-import './../../css/Home.css';
+import '../../css/Home.css';
 
 const BasicInstructions = props => {
   return (
