+++ frontend/src/components/Nav.js
@@ -1,9 +1,9 @@
 import React from "react";
 import MediaQuery from "react-responsive";
 import { NavLink } from "react-router-dom";
-import "./../css/Nav.css";
-import srLogo from "./../images/sr-logo-sm.png";
-import rcrLogo from "./../images/rcr-logo.png";
+import "../css/Nav.css";
+import srLogo from "../images/sr-logo-sm.png";
+import rcrLogo from "../images/rcr-logo.png";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { CSSTransition } from "react-transition-group";
