+++ frontend/src/components/Results/Results.js
@@ -7,7 +7,7 @@ import SearchBar from '../Home/SearchBar';
 import Cards from './Cards';
 import SimpleMap from './Map';
 import { getFilteredRecords, detailsQueryBuilder } from '../../utils/api';
-import './../../css/Results.css';
+import '../../css/Results.css';
 
 class Results extends React.Component {
   state = {
